Add multiprocessing as an option with the pickle_safe argument

+++ keras/engine/training.py
@@ -6,6 +6,11 @@ import copy
 import time
 import numpy as np
 import multiprocessing
+import threading
+try:
+    import queue
+except ImportError:
+    import Queue as queue
 
 from .topology import Container
 from .. import backend as K
@@ -391,18 +396,20 @@ def standardize_weights(y, sample_weight=None, class_weight=None,
 
 
 def generator_queue(generator, max_q_size=10,
-                    wait_time=0.05, nb_worker=2):
-    '''Builds a multiprocessing queue out of a data generator.
+                    wait_time=0.05, nb_worker=1, pickle_safe=False):
+    '''Builds a queue out of a data generator.
+    If pickle_safe, use a multiprocessing approach. Else, use threading.
     Used in `fit_generator`, `evaluate_generator`, `predict_generator`.
-    Note that because this implementation relies on multiprocessing,
-    the generator should not depend on non picklable arguments
-    which can't be passed easily to subprocesses.
 
     '''
 
-    processes = []
-    q = multiprocessing.Queue(maxsize=max_q_size)
-    _stop = multiprocessing.Event()
+    generator_threads = []
+    if pickle_safe:
+        q = multiprocessing.Queue(maxsize=max_q_size)
+        _stop = multiprocessing.Event()
+    else:
+        q = queue.Queue()
+        _stop = threading.Event()
 
     try:
 
@@ -422,24 +429,27 @@ def generator_queue(generator, max_q_size=10,
                     raise
 
         for i in range(nb_worker):
-            # Reset random seed
-            np.random.seed()
-
-            proc = multiprocessing.Process(target=data_generator_task)
-            processes.append(proc)
-            proc.daemon = True
-            proc.start()
+            if pickle_safe:
+                # Reset random seed else all children processes share the same seed
+                np.random.seed()
+                thread = multiprocessing.Process(target=data_generator_task)
+            else:
+                thread = threading.Thread(target=data_generator_task)
+            generator_threads.append(thread)
+            thread.daemon = True
+            thread.start()
 
     except:
-        # Terminate all daemon processes
         _stop.set()
-        for p in processes:
-            if p.is_alive():
-                p.terminate()
-        q.close()
+        if pickle_safe:
+            # Terminate all daemon processes
+            for p in generator_threads:
+                if p.is_alive():
+                    p.terminate()
+            q.close()
         raise
 
-    return q, _stop, processes
+    return q, _stop
 
 
 class Model(Container):
@@ -1268,7 +1278,7 @@ class Model(Container):
     def fit_generator(self, generator, samples_per_epoch, nb_epoch,
                       verbose=1, callbacks=[],
                       validation_data=None, nb_val_samples=None,
-                      class_weight={}, max_q_size=10, nb_worker=2):
+                      class_weight={}, max_q_size=10, nb_worker=1, pickle_safe=False):
         '''Fits the model on data generated batch-by-batch by
         a Python generator.
         The generator is run in parallel to the model, for efficiency.
@@ -1284,9 +1294,6 @@ class Model(Container):
                 The generator is expected to loop over its data
                 indefinitely. An epoch finishes when `samples_per_epoch`
                 samples have been seen by the model.
-                Note that because this implementation relies on multiprocessing,
-                the generator should not depend on non picklable arguments
-                which can't be passed easily to subprocesses.
             samples_per_epoch: integer, number of samples to process before
                 going to the next epoch.
             nb_epoch: integer, total number of iterations on the data.
@@ -1302,7 +1309,11 @@ class Model(Container):
             class_weight: dictionary mapping class indices to a weight
                 for the class.
             max_q_size: maximum size for the generator queue
-            nb_worker: maximum number of processes to spin up
+            nb_worker: maximum number of processes to spin up when using process based threading
+            pickle_safe: if True, use process based threading. Note that because
+                this implementation relies on multiprocessing, you should not pass non
+                non picklable arguments to the generator as they can't be passed
+                easily to children processes.
 
         # Returns
             A `History` object.
@@ -1381,7 +1392,8 @@ class Model(Container):
             self.validation_data = None
 
         # start generator thread storing batches into a queue
-        data_gen_queue, _stop, processes = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker)
+        data_gen_queue, _stop = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker,
+                                                pickle_safe=pickle_safe)
 
         callback_model.stop_training = False
         while epoch < nb_epoch:
@@ -1474,13 +1486,12 @@ class Model(Container):
                 break
 
         _stop.set()
-        for p in processes:
-            p.terminate()
-        data_gen_queue.close()
+        if pickle_safe:
+            data_gen_queue.close()
         callbacks.on_train_end()
         return self.history
 
-    def evaluate_generator(self, generator, val_samples, max_q_size=10, nb_worker=2):
+    def evaluate_generator(self, generator, val_samples, max_q_size=10, nb_worker=1, pickle_safe=False):
         '''Evaluates the model on a data generator. The generator should
         return the same kind of data as accepted by `test_on_batch`.
 
@@ -1488,14 +1499,15 @@ class Model(Container):
             generator:
                 generator yielding tuples (inputs, targets)
                 or (inputs, targets, sample_weights)
-                    Note that because this implementation relies on multiprocessing,
-                    the generator should not depend on non picklable arguments
-                    which can't be passed easily to subprocesses.
             val_samples:
                 total number of samples to generate from `generator`
                 before returning.
             max_q_size: maximum size for the generator queue
-            nb_worker: maximum number of processes to spin up
+            nb_worker: maximum number of processes to spin up when using process based threading
+            pickle_safe: if True, use process based threading. Note that because
+                this implementation relies on multiprocessing, you should not pass non
+                non picklable arguments to the generator as they can't be passed
+                easily to children processes.
 
         # Returns
             Scalar test loss (if the model has a single output and no metrics)
@@ -1509,7 +1521,8 @@ class Model(Container):
         wait_time = 0.01
         all_outs = []
         weights = []
-        data_gen_queue, _stop, processes = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker)
+        data_gen_queue, _stop = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker,
+                                                pickle_safe=pickle_safe)
 
         while processed_samples < val_samples:
             generator_output = None
@@ -1553,9 +1566,8 @@ class Model(Container):
             weights.append(nb_samples)
 
         _stop.set()
-        for p in processes:
-            p.terminate()
-        data_gen_queue.close()
+        if pickle_safe:
+            data_gen_queue.close()
         if type(outs) is not list:
             return np.average(np.asarray(all_outs),
                               weights=weights)
@@ -1566,21 +1578,21 @@ class Model(Container):
                                            weights=weights))
             return averages
 
-    def predict_generator(self, generator, val_samples, max_q_size=10, nb_worker=2):
+    def predict_generator(self, generator, val_samples, max_q_size=10, nb_worker=1, pickle_safe=False):
         '''Generates predictions for the input samples from a data generator.
         The generator should return the same kind of data as accepted by
         `predict_on_batch`.
 
         # Arguments
             generator: generator yielding batches of input samples.
-                Note that because this implementation relies on multiprocessing,
-                the generator should not depend on non picklable arguments
-                which can't be passed easily to subprocesses.
             val_samples: total number of samples to generate from `generator`
                 before returning.
             max_q_size: maximum size for the generator queue
-            nb_worker: maximum number of processes to spin up
-
+            nb_worker: maximum number of processes to spin up when using process based threading
+            pickle_safe: if True, use process based threading. Note that because
+                this implementation relies on multiprocessing, you should not pass non
+                non picklable arguments to the generator as they can't be passed
+                easily to children processes.
 
         # Returns
             Numpy array(s) of predictions.
@@ -1590,7 +1602,8 @@ class Model(Container):
         processed_samples = 0
         wait_time = 0.01
         all_outs = []
-        data_gen_queue, _stop, processes = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker)
+        data_gen_queue, _stop = generator_queue(generator, max_q_size=max_q_size, nb_worker=nb_worker,
+                                                pickle_safe=pickle_safe)
 
         while processed_samples < val_samples:
             generator_output = None
@@ -1642,9 +1655,8 @@ class Model(Container):
             processed_samples += nb_samples
 
         _stop.set()
-        for p in processes:
-            p.terminate()
-        data_gen_queue.close()
+        if pickle_safe:
+            data_gen_queue.close()
         if len(all_outs) == 1:
             return all_outs[0]
         return all_outs
